fix: rename package to encryptionatrest
