+++ pkg/handlers/generic/mutation/encryptionatrest/inject_test.go
@@ -1,7 +1,7 @@
 // Copyright 2024 Nutanix. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package encryption
+package encryptionatrest
 
 import (
 	"testing"
